WIP: fix hardhat config to properly compile all
It builds, but now I notice that the OracleRewardShadowVault is too big in size. I need to fix that. I moved the RamsesV3Pool and the NonfungiblePositionManager files to the reference folder as we don't actually want to build them (it's just to help understand these contracts).

+++ contracts-shadow/src/interfaces/IOracleRewardShadowVault.sol
@@ -59,7 +59,7 @@ interface IOracleRewardShadowVault is IMinimalVault, IERC20Upgradeable {
 
     // Structs
     struct QueuedWithdrawal {
-        uint128 totalQueuedShares;
+        uint256 totalQueuedShares;
         uint128 totalAmountX;
         uint128 totalAmountY;
         mapping(address => uint256) userWithdrawals;
